End of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -59,12 +60,13 @@ public class AuctionsController : ControllerBase
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreatAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        //TODO add current user as seller
-        auction.Seller = "test";
+        
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
         
@@ -82,6 +84,7 @@ public class AuctionsController : ControllerBase
 
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id,
                                                     UpdateAuctionDto updateAuctionDto)
@@ -91,7 +94,7 @@ public class AuctionsController : ControllerBase
 
         if(auction == null) return NotFound();
 
-        //TODO: check seller == username
+        if(auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -107,13 +110,16 @@ public class AuctionsController : ControllerBase
         return BadRequest("Problem saving changes");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
         var auction = await _context.Auctions.FindAsync(id);
 
         if(auction == null) return NotFound();
-        //TODO check seller == username
+        
+        if(auction.Seller != User.Identity.Name) return Forbid();
+
          _context.Auctions.Remove(auction);
 
         await _publishEndpoint.Publish<AuctionDeleted>( new { Id = auction.Id.ToString() });
